+++ Micro.Auth.Api/GraphQL/Mutation.cs
@@ -2,6 +2,8 @@ using GraphQL;
 using GraphQL.Types;
 using Micro.Auth.Api.GraphQL.Inputs;
 using Micro.Auth.Api.GraphQL.Types;
+using Micro.Auth.Business.EmailVerification;
+using Micro.Auth.Business.PasswordManager;
 using Micro.Auth.Business.Users;
 using Micro.Auth.Business.Users.ViewModels;
 
@@ -9,7 +11,7 @@ namespace Micro.Auth.Api.GraphQL
 {
     public class Mutation : ObjectGraphType
     {
-        public Mutation(IUserService userService)
+        public Mutation(IUserService userService, IPasswordManager passwordManager, IEmailVerificationService verification)
         {
             FieldAsync<NonNullGraphType<UserType>, User>("register",
                 arguments: new QueryArguments(RegisterInputType.BuildArgument()),
@@ -17,24 +19,24 @@ namespace Micro.Auth.Api.GraphQL
 
             FieldAsync<NonNullGraphType<UserType>, User>("verifyEmail",
                 arguments: new QueryArguments(VerifyEmailInputType.BuildArgument()),
-                resolve: x => userService.ConfirmEmail(x.GetArgument<VerifyEmailInput>("input")));
+                resolve: x => verification.ConfirmEmail(x.GetArgument<VerifyEmailInput>("input")));
 
             FieldAsync<NonNullGraphType<ResultType>, Result>("sendActivationEmail",
                 arguments: new QueryArguments(new QueryArgument<NonNullGraphType<StringGraphType>> {Name = "login"}),
-                resolve: x => userService.SendActivationEmail(x.GetArgument<string>("login")));
+                resolve: x => verification.SendActivationEmail(x.GetArgument<string>("login")));
 
             FieldAsync<NonNullGraphType<ResultType>, Result>("requestPasswordReset",
                 arguments: new QueryArguments(new QueryArgument<NonNullGraphType<StringGraphType>> {Name = "login"}),
-                resolve: x => userService.RequestPasswordReset(x.GetArgument<string>("login")));
+                resolve: x => passwordManager.RequestPasswordReset(x.GetArgument<string>("login")));
 
             // todo: find a way to get UserId
             FieldAsync<NonNullGraphType<UserType>, User>("changePassword",
                 arguments: new QueryArguments(ChangePasswordInput.BuildArgument()),
-                resolve: x => userService.ChangePassword("/////----/////", x.GetArgument<ChangePasswordRequest>("input")));
+                resolve: x => passwordManager.ChangePassword("/////----/////", x.GetArgument<ChangePasswordRequest>("input")));
 
             FieldAsync<NonNullGraphType<UserType>, User>("resetPassword",
                 arguments: new QueryArguments(ResetPasswordInput.BuildArgument()),
-                resolve: x => userService.ResetPassword(x.GetArgument<ResetPasswordRequest>("input")));
+                resolve: x => passwordManager.ResetPassword(x.GetArgument<ResetPasswordRequest>("input")));
         }
     }
 }
