+++ Micro.Auth.Api/GraphQL/Query.cs
@@ -1,6 +1,7 @@
 using GraphQL;
 using GraphQL.Types;
 using Micro.Auth.Api.GraphQL.Types;
+using Micro.Auth.Business.Availability;
 using Micro.Auth.Business.Users;
 using Micro.Auth.Business.Users.ViewModels;
 
@@ -8,7 +9,7 @@ namespace Micro.Auth.Api.GraphQL
 {
     public class Query : ObjectGraphType
     {
-        public Query(IUserService userService)
+        public Query(IUserService userService, IAvailabilityService availabilityService)
         {
             FieldAsync<NonNullGraphType<UserType>, User>("user",
                 arguments: new QueryArguments(new QueryArgument<NonNullGraphType<StringGraphType>> {Name = "id"}),
@@ -28,11 +29,11 @@ namespace Micro.Auth.Api.GraphQL
 
             FieldAsync<NonNullGraphType<AvailabilityResultType>, AvailabilityResponse>("availabilityByUsername",
                 arguments: new QueryArguments(new QueryArgument<NonNullGraphType<StringGraphType>> {Name = "username"}),
-                resolve: x => userService.AvailabilityByLogin(x.GetArgument<string>("username")));
+                resolve: x => availabilityService.AvailabilityByLogin(x.GetArgument<string>("username")));
 
             FieldAsync<NonNullGraphType<AvailabilityResultType>, AvailabilityResponse>("availabilityByEmail",
                 arguments: new QueryArguments(new QueryArgument<NonNullGraphType<StringGraphType>> {Name = "email"}),
-                resolve: x => userService.AvailabilityByEmail(x.GetArgument<string>("email")));
+                resolve: x => availabilityService.AvailabilityByEmail(x.GetArgument<string>("email")));
         }
     }
 }
