+++ Micro.Auth.Business/Availability/IAvailabilityService.cs
@@ -1,0 +1,11 @@
+using System.Threading.Tasks;
+using Micro.Auth.Business.Users.ViewModels;
+
+namespace Micro.Auth.Business.Availability
+{
+    public interface IAvailabilityService
+    {
+        Task<AvailabilityResponse> AvailabilityByEmail(string email);
+        Task<AvailabilityResponse> AvailabilityByLogin(string login);
+    }
+}
